Use the real Knative Service URL instead of assuming example.com
This plumbs through the Knative Service URLs returned from
https://github.com/serverless-components/knative-serving/pull/2 into
the CLI. This change depends on that
serverless-components/knative-serving one to get merged first and the
version bumped in package.json to provide the new Knative Service URLs.

This fixes #3.

It partially addresses #2 as well, but the actual function invocation
logic still tries the Istio IP if one is found. If not, it falls back
to directly using the Knative Service URL given.